refactor(api): modularize project
