+++ apps/api/Cargo.toml
@@ -17,7 +17,12 @@ rand_core = { version = "0.6.4", features = ["std"] }
 redis = { version = "0.23.0", features = ["tokio-comp"] }
 serde = { version = "1.0.159", features = ["derive"] }
 serde_json = "1.0.95"
-sqlx = { version = "0.6.3", features = ["runtime-async-std-native-tls", "postgres", "chrono", "uuid"] }
+sqlx = { version = "0.6.3", features = [
+  "runtime-async-std-native-tls",
+  "postgres",
+  "chrono",
+  "uuid",
+] }
 time = "0.3.20"
 tokio = { version = "1.27.0", features = ["full"] }
 tower-http = { version = "0.4.0", features = ["cors"] }
