+++ apps/api/src/handler.rs
@@ -1,0 +1,16 @@
+use std::sync::Arc;
+
+use argon2::{password_hash::SaltString, Argon2, PasswordHash, PasswordHasher, PasswordVerifier};
+use axum::{
+    extract::State,
+    http::{header, HeaderMap, Response, StatusCode},
+    response::IntoResponse,
+    Extension, Json,
+};
+use axum_extra::extract::{
+    cookie::{Cookie, SameSite},
+    CookieJar,
+};
+use rand_core::OsRng;
+use redis::AsyncCommands;
+use serde_json::json;
