+++ apps/api/src/lib.rs
@@ -1,10 +1,75 @@
+use std::sync::Arc;
+
+use axum::http::{
+    header::{ACCEPT, AUTHORIZATION, CONTENT_TYPE},
+    HeaderValue, Method,
+};
+use redis::Client;
+use sqlx::{postgres::PgPoolOptions, Pool, Postgres};
+use tower_http::cors::CorsLayer;
+
+pub mod model;
+pub mod response;
 mod routes;
+pub mod utils;
 
-use routes::create_routes;
+use routes::create_router;
+use utils::config::Config;
 
-pub async fn start_http_server() {
-    // Build our application with the routes we've defined
-    let app = create_routes();
+pub struct AppState {
+    db: Pool<Postgres>,
+    env: Config,
+    redis_client: Client,
+}
+
+pub async fn start_http_server(config: Config) {
+    let pool = match PgPoolOptions::new()
+        .max_connections(10)
+        .connect(&config.database_url)
+        .await
+    {
+        Ok(pool) => {
+            println!("✅ Connection to the database is successful!");
+            pool
+        }
+        Err(err) => {
+            println!("❌ Failed to connect to the database: {:?}", err);
+            std::process::exit(1);
+        }
+    };
+
+    let redis_client = match Client::open(config.redis_url.to_owned()) {
+        Ok(client) => {
+            println!("✅ Connection to the redis is successful!");
+            client
+        }
+        Err(e) => {
+            println!("❌ Error connecting to Redis: {}", e);
+            std::process::exit(1);
+        }
+    };
+
+    let cors = CorsLayer::new()
+        .allow_origin("http://localhost:3000".parse::<HeaderValue>().unwrap())
+        .allow_methods([
+            Method::GET,
+            Method::POST,
+            Method::PUT,
+            Method::PATCH,
+            Method::DELETE,
+        ])
+        .allow_credentials(true)
+        .allow_headers([AUTHORIZATION, ACCEPT, CONTENT_TYPE]);
+
+    // Build the application with the defined routes
+    let app = create_router(Arc::new(AppState {
+        db: pool.clone(),
+        env: config.clone(),
+        redis_client: redis_client.clone(),
+    }))
+    .layer(cors);
+
+    println!("🚀 API server started successfully");
 
     // Run it with hyper on localhost:5000
     axum::Server::bind(&"0.0.0.0:5000".parse().unwrap())
