+++ apps/api/src/main.rs
@@ -1,29 +1,6 @@
-mod config;
-mod handler;
-mod jwt_auth;
-mod model;
-mod response;
-mod route;
-mod token;
+use api::{start_http_server, utils::config::Config};
 
-use config::Config;
-use std::sync::Arc;
-
-use axum::http::{
-    header::{ACCEPT, AUTHORIZATION, CONTENT_TYPE},
-    HeaderValue, Method,
-};
 use dotenv::dotenv;
-use redis::Client;
-use route::create_router;
-use sqlx::{postgres::PgPoolOptions, Pool, Postgres};
-use tower_http::cors::CorsLayer;
-
-pub struct AppState {
-    db: Pool<Postgres>,
-    env: Config,
-    redis_client: Client,
-}
 
 #[tokio::main]
 async fn main() {
@@ -31,48 +8,5 @@ async fn main() {
 
     let config = Config::init();
 
-    let pool = match PgPoolOptions::new()
-        .max_connections(10)
-        .connect(&config.database_url)
-        .await
-    {
-        Ok(pool) => {
-            println!("✅Connection to the database is successful!");
-            pool
-        }
-        Err(err) => {
-            println!("🔥 Failed to connect to the database: {:?}", err);
-            std::process::exit(1);
-        }
-    };
-
-    let redis_client = match Client::open(config.redis_url.to_owned()) {
-        Ok(client) => {
-            println!("✅Connection to the redis is successful!");
-            client
-        }
-        Err(e) => {
-            println!("🔥 Error connecting to Redis: {}", e);
-            std::process::exit(1);
-        }
-    };
-
-    let cors = CorsLayer::new()
-        .allow_origin("http://localhost:3000".parse::<HeaderValue>().unwrap())
-        .allow_methods([Method::GET, Method::POST, Method::PATCH, Method::DELETE])
-        .allow_credentials(true)
-        .allow_headers([AUTHORIZATION, ACCEPT, CONTENT_TYPE]);
-
-    let app = create_router(Arc::new(AppState {
-        db: pool.clone(),
-        env: config.clone(),
-        redis_client: redis_client.clone(),
-    }))
-    .layer(cors);
-
-    println!("🚀 Server started successfully");
-    axum::Server::bind(&"0.0.0.0:5000".parse().unwrap())
-        .serve(app.into_make_service())
-        .await
-        .unwrap();
+    start_http_server(config).await;
 }
