+++ apps/api/src/routes/auth/login.rs
@@ -1,0 +1,125 @@
+use std::sync::Arc;
+
+use argon2::{Argon2, PasswordHash, PasswordVerifier};
+use axum::{
+    extract::State,
+    http::{header, HeaderMap, Response, StatusCode},
+    response::IntoResponse,
+    Json,
+};
+use axum_extra::extract::cookie::{Cookie, SameSite};
+use serde_json::json;
+
+use crate::{
+    model::{LoginUserSchema, User},
+    utils::{generate_token, save_token_data_to_redis},
+    AppState,
+};
+
+pub async fn login_user_handler(
+    State(data): State<Arc<AppState>>,
+    Json(body): Json<LoginUserSchema>,
+) -> Result<impl IntoResponse, (StatusCode, Json<serde_json::Value>)> {
+    let user = sqlx::query_as!(
+        User,
+        "SELECT * FROM users WHERE email = $1",
+        body.email.to_ascii_lowercase()
+    )
+    .fetch_optional(&data.db)
+    .await
+    .map_err(|e| {
+        let error_response = serde_json::json!({
+            "status": "error",
+            "message": format!("Database error: {}", e),
+        });
+        (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
+    })?
+    .ok_or_else(|| {
+        let error_response = serde_json::json!({
+            "status": "fail",
+            "message": "Invalid email or password",
+        });
+        (StatusCode::BAD_REQUEST, Json(error_response))
+    })?;
+
+    let is_valid = match PasswordHash::new(&user.password) {
+        Ok(parsed_hash) => Argon2::default()
+            .verify_password(body.password.as_bytes(), &parsed_hash)
+            .map_or(false, |_| true),
+        Err(_) => false,
+    };
+
+    if !is_valid {
+        let error_response = serde_json::json!({
+            "status": "fail",
+            "message": "Invalid email or password"
+        });
+        return Err((StatusCode::BAD_REQUEST, Json(error_response)));
+    }
+
+    let access_token_details = generate_token(
+        user.id,
+        data.env.access_token_max_age,
+        data.env.access_token_private_key.to_owned(),
+    )?;
+    let refresh_token_details = generate_token(
+        user.id,
+        data.env.refresh_token_max_age,
+        data.env.refresh_token_private_key.to_owned(),
+    )?;
+
+    save_token_data_to_redis(&data, &access_token_details, data.env.access_token_max_age).await?;
+    save_token_data_to_redis(
+        &data,
+        &refresh_token_details,
+        data.env.refresh_token_max_age,
+    )
+    .await?;
+
+    let access_cookie = Cookie::build(
+        "access_token",
+        access_token_details.token.clone().unwrap_or_default(),
+    )
+    .path("/")
+    .max_age(time::Duration::minutes(data.env.access_token_max_age * 60))
+    .same_site(SameSite::Lax)
+    .http_only(true)
+    .finish();
+    let refresh_cookie = Cookie::build(
+        "refresh_token",
+        refresh_token_details.token.unwrap_or_default(),
+    )
+    .path("/")
+    .max_age(time::Duration::minutes(data.env.refresh_token_max_age * 60))
+    .same_site(SameSite::Lax)
+    .http_only(true)
+    .finish();
+
+    let logged_in_cookie = Cookie::build("logged_in", "true")
+        .path("/")
+        .max_age(time::Duration::minutes(data.env.access_token_max_age * 60))
+        .same_site(SameSite::Lax)
+        .http_only(false)
+        .finish();
+
+    let mut response = Response::new(
+        json!({"status": "success", "access_token": access_token_details.token.unwrap()})
+            .to_string(),
+    );
+    let mut headers = HeaderMap::new();
+    headers.append(
+        header::SET_COOKIE,
+        access_cookie.to_string().parse().unwrap(),
+    );
+    headers.append(
+        header::SET_COOKIE,
+        refresh_cookie.to_string().parse().unwrap(),
+    );
+    headers.append(
+        header::SET_COOKIE,
+        logged_in_cookie.to_string().parse().unwrap(),
+    );
+
+    response.headers_mut().extend(headers);
+    Ok(response)
+}
