+++ apps/api/src/routes/auth/logout.rs
@@ -1,0 +1,115 @@
+use std::sync::Arc;
+
+use axum::{
+    extract::State,
+    http::{header, HeaderMap, Response, StatusCode},
+    response::IntoResponse,
+    Extension, Json,
+};
+use axum_extra::extract::{
+    cookie::{Cookie, SameSite},
+    CookieJar,
+};
+use redis::AsyncCommands;
+use serde_json::json;
+
+use crate::{
+    utils::{jwt_auth::JWTAuthMiddleware, token},
+    AppState,
+};
+
+pub async fn logout_handler(
+    cookie_jar: CookieJar,
+    Extension(auth_guard): Extension<JWTAuthMiddleware>,
+    State(data): State<Arc<AppState>>,
+) -> Result<impl IntoResponse, (StatusCode, Json<serde_json::Value>)> {
+    let message = "Token is invalid or session has expired";
+
+    let refresh_token = cookie_jar
+        .get("refresh_token")
+        .map(|cookie| cookie.value().to_string())
+        .ok_or_else(|| {
+            let error_response = serde_json::json!({
+                "status": "fail",
+                "message": message
+            });
+            (StatusCode::FORBIDDEN, Json(error_response))
+        })?;
+
+    let refresh_token_details =
+        match token::verify_jwt_token(data.env.refresh_token_public_key.to_owned(), &refresh_token)
+        {
+            Ok(token_details) => token_details,
+            Err(e) => {
+                let error_response = serde_json::json!({
+                    "status": "fail",
+                    "message": format_args!("{:?}", e)
+                });
+                return Err((StatusCode::UNAUTHORIZED, Json(error_response)));
+            }
+        };
+
+    let mut redis_client = data
+        .redis_client
+        .get_async_connection()
+        .await
+        .map_err(|e| {
+            let error_response = serde_json::json!({
+                "status": "error",
+                "message": format!("Redis error: {}", e),
+            });
+            (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
+        })?;
+
+    redis_client
+        .del(&[
+            refresh_token_details.token_uuid.to_string(),
+            auth_guard.access_token_uuid.to_string(),
+        ])
+        .await
+        .map_err(|e| {
+            let error_response = serde_json::json!({
+                "status": "error",
+                "message": format_args!("{:?}", e)
+            });
+            (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
+        })?;
+
+    let access_cookie = Cookie::build("access_token", "")
+        .path("/")
+        .max_age(time::Duration::minutes(-1))
+        .same_site(SameSite::Lax)
+        .http_only(true)
+        .finish();
+    let refresh_cookie = Cookie::build("refresh_token", "")
+        .path("/")
+        .max_age(time::Duration::minutes(-1))
+        .same_site(SameSite::Lax)
+        .http_only(true)
+        .finish();
+
+    let logged_in_cookie = Cookie::build("logged_in", "true")
+        .path("/")
+        .max_age(time::Duration::minutes(-1))
+        .same_site(SameSite::Lax)
+        .http_only(false)
+        .finish();
+
+    let mut headers = HeaderMap::new();
+    headers.append(
+        header::SET_COOKIE,
+        access_cookie.to_string().parse().unwrap(),
+    );
+    headers.append(
+        header::SET_COOKIE,
+        refresh_cookie.to_string().parse().unwrap(),
+    );
+    headers.append(
+        header::SET_COOKIE,
+        logged_in_cookie.to_string().parse().unwrap(),
+    );
+
+    let mut response = Response::new(json!({"status": "success"}).to_string());
+    response.headers_mut().extend(headers);
+    Ok(response)
+}
