+++ apps/api/src/routes/auth/mod.rs
@@ -1,0 +1,4 @@
+pub mod login;
+pub mod logout;
+pub mod refresh;
+pub mod register;
