+++ apps/api/src/routes/auth/refresh.rs
@@ -1,0 +1,143 @@
+use std::sync::Arc;
+
+use axum::{
+    extract::State,
+    http::{header, HeaderMap, Response, StatusCode},
+    response::IntoResponse,
+    Json,
+};
+use axum_extra::extract::{
+    cookie::{Cookie, SameSite},
+    CookieJar,
+};
+use redis::AsyncCommands;
+use serde_json::json;
+
+use crate::{
+    model::User,
+    utils::{generate_token, save_token_data_to_redis, token},
+    AppState,
+};
+
+pub async fn refresh_access_token_handler(
+    cookie_jar: CookieJar,
+    State(data): State<Arc<AppState>>,
+) -> Result<impl IntoResponse, (StatusCode, Json<serde_json::Value>)> {
+    let message = "could not refresh access token";
+
+    let refresh_token = cookie_jar
+        .get("refresh_token")
+        .map(|cookie| cookie.value().to_string())
+        .ok_or_else(|| {
+            let error_response = serde_json::json!({
+                "status": "fail",
+                "message": message
+            });
+            (StatusCode::FORBIDDEN, Json(error_response))
+        })?;
+
+    let refresh_token_details =
+        match token::verify_jwt_token(data.env.refresh_token_public_key.to_owned(), &refresh_token)
+        {
+            Ok(token_details) => token_details,
+            Err(e) => {
+                let error_response = serde_json::json!({
+                    "status": "fail",
+                    "message": format_args!("{:?}", e)
+                });
+                return Err((StatusCode::UNAUTHORIZED, Json(error_response)));
+            }
+        };
+
+    let mut redis_client = data
+        .redis_client
+        .get_async_connection()
+        .await
+        .map_err(|e| {
+            let error_response = serde_json::json!({
+                "status": "error",
+                "message": format!("Redis error: {}", e),
+            });
+            (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
+        })?;
+
+    let redis_token_user_id = redis_client
+        .get::<_, String>(refresh_token_details.token_uuid.to_string())
+        .await
+        .map_err(|_| {
+            let error_response = serde_json::json!({
+                "status": "error",
+                "message": "Token is invalid or session has expired",
+            });
+            (StatusCode::UNAUTHORIZED, Json(error_response))
+        })?;
+
+    let user_id_uuid = uuid::Uuid::parse_str(&redis_token_user_id).map_err(|_| {
+        let error_response = serde_json::json!({
+            "status": "error",
+            "message": "Token is invalid or session has expired",
+        });
+        (StatusCode::UNAUTHORIZED, Json(error_response))
+    })?;
+
+    let user = sqlx::query_as!(User, "SELECT * FROM users WHERE id = $1", user_id_uuid)
+        .fetch_optional(&data.db)
+        .await
+        .map_err(|e| {
+            let error_response = serde_json::json!({
+                "status": "fail",
+                "message": format!("Error fetching user from database: {}", e),
+            });
+            (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
+        })?;
+
+    let user = user.ok_or_else(|| {
+        let error_response = serde_json::json!({
+            "status": "fail",
+            "message": "The user belonging to this token no longer exists".to_string(),
+        });
+        (StatusCode::UNAUTHORIZED, Json(error_response))
+    })?;
+
+    let access_token_details = generate_token(
+        user.id,
+        data.env.access_token_max_age,
+        data.env.access_token_private_key.to_owned(),
+    )?;
+
+    save_token_data_to_redis(&data, &access_token_details, data.env.access_token_max_age).await?;
+
+    let access_cookie = Cookie::build(
+        "access_token",
+        access_token_details.token.clone().unwrap_or_default(),
+    )
+    .path("/")
+    .max_age(time::Duration::minutes(data.env.access_token_max_age * 60))
+    .same_site(SameSite::Lax)
+    .http_only(true)
+    .finish();
+
+    let logged_in_cookie = Cookie::build("logged_in", "true")
+        .path("/")
+        .max_age(time::Duration::minutes(data.env.access_token_max_age * 60))
+        .same_site(SameSite::Lax)
+        .http_only(false)
+        .finish();
+
+    let mut response = Response::new(
+        json!({"status": "success", "access_token": access_token_details.token.unwrap()})
+            .to_string(),
+    );
+    let mut headers = HeaderMap::new();
+    headers.append(
+        header::SET_COOKIE,
+        access_cookie.to_string().parse().unwrap(),
+    );
+    headers.append(
+        header::SET_COOKIE,
+        logged_in_cookie.to_string().parse().unwrap(),
+    );
+
+    response.headers_mut().extend(headers);
+    Ok(response)
+}
