+++ apps/api/src/routes/auth/register.rs
@@ -1,0 +1,74 @@
+use std::sync::Arc;
+
+use argon2::{password_hash::SaltString, Argon2, PasswordHasher};
+use axum::{extract::State, http::StatusCode, response::IntoResponse, Json};
+use rand_core::OsRng;
+
+use crate::{
+    model::{RegisterUserSchema, User},
+    utils::user::filter_user_record,
+    AppState,
+};
+
+pub async fn register_user_handler(
+    State(data): State<Arc<AppState>>,
+    Json(body): Json<RegisterUserSchema>,
+) -> Result<impl IntoResponse, (StatusCode, Json<serde_json::Value>)> {
+    let user_exists: Option<bool> =
+        sqlx::query_scalar("SELECT EXISTS(SELECT 1 FROM users WHERE email = $1)")
+            .bind(body.email.to_owned().to_ascii_lowercase())
+            .fetch_one(&data.db)
+            .await
+            .map_err(|e| {
+                let error_response = serde_json::json!({
+                    "status": "fail",
+                    "message": format!("Database error: {}", e),
+                });
+                (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
+            })?;
+
+    if let Some(exists) = user_exists {
+        if exists {
+            let error_response = serde_json::json!({
+                "status": "fail",
+                "message": "User with that email already exists",
+            });
+            return Err((StatusCode::CONFLICT, Json(error_response)));
+        }
+    }
+
+    let salt = SaltString::generate(&mut OsRng);
+    let hashed_password = Argon2::default()
+        .hash_password(body.password.as_bytes(), &salt)
+        .map_err(|e| {
+            let error_response = serde_json::json!({
+                "status": "fail",
+                "message": format!("Error while hashing password: {}", e),
+            });
+            (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
+        })
+        .map(|hash| hash.to_string())?;
+
+    let user = sqlx::query_as!(
+        User,
+        "INSERT INTO users (name,email,password) VALUES ($1, $2, $3) RETURNING *",
+        body.name.to_string(),
+        body.email.to_string().to_ascii_lowercase(),
+        hashed_password
+    )
+    .fetch_one(&data.db)
+    .await
+    .map_err(|e| {
+        let error_response = serde_json::json!({
+            "status": "fail",
+            "message": format!("Database error: {}", e),
+        });
+        (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
+    })?;
+
+    let user_response = serde_json::json!({"status": "success","data": serde_json::json!({
+        "user": filter_user_record(&user)
+    })});
+
+    Ok(Json(user_response))
+}
