+++ apps/api/src/routes/info/mod.rs
@@ -1,0 +1,1 @@
+pub mod healthchecker;
