+++ apps/api/src/routes/users/mod.rs
@@ -1,0 +1,1 @@
+pub mod get_me;
