+++ apps/api/src/utils/user.rs
@@ -1,0 +1,14 @@
+use crate::{model::User, response::FilteredUser};
+
+pub fn filter_user_record(user: &User) -> FilteredUser {
+    FilteredUser {
+        id: user.id.to_string(),
+        email: user.email.to_owned(),
+        name: user.name.to_owned(),
+        photo: user.photo.to_owned(),
+        role: user.role.to_owned(),
+        verified: user.verified,
+        createdAt: user.created_at.unwrap(),
+        updatedAt: user.updated_at.unwrap(),
+    }
+}
